requirements.txt
numpy
sentence-transformers
pdfplumber
python-docx
faiss-cpu